Handle multiple options with same name correctly
If the interface should have several entries with the same name,
only the last one in the hash made it into the config file. This
change enables the value to be an array which is then expanded
to multiple lines. Now one can write
'up' => [ 'ip route add 10.0.0.0/8 via 10.0.0.1', 'ip route add 192.168.0.0/16 via 10.0.0.1' ]
and have multiple up commands in the interface.